Changed gameStage.html to represent a universal board, linked UserAuthPage.html to index.html, fixed linter issue, changed the nameing conventions so all java files are capital and others are lowercase
