+++ src/main/java/com/google/sps/data/Player.java
@@ -12,57 +12,57 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import com.google.appengine.api.datastore.Key;
+package com.google.sps.data;
 
 public class Player {
-    private String displayName;
-    private String email;
-    private String id;
-    private String imageID;
-    private String currentPageID;
+  private String displayName;
+  private String email;
+  private String id;
+  private String imageID;
+  private String currentPageID;
 
-    public Player (String displayName, String email){
-        this.displayName = displayName;
-        this.email = email;
-    }
+  public Player(String displayName, String email) {
+    this.displayName = displayName;
+    this.email = email;
+  }
 
-    public String getdisplayName() {
-        return displayName;
-    }
+  public String getDisplayName() {
+    return displayName;
+  }
 
-    public String getEmail() {
-        return email;
-    }
+  public String getEmail() {
+    return email;
+  }
 
-    public String getID(){
-        return id;
-    }
+  public String getID(){
+    return id;
+  }
 
-    public String getImageID() {
-        return imageID;
-    }
+  public String getImageID() {
+    return imageID;
+  }
 
-    public String getCurrentPageID() {
-        return currentPageID;
-    }
+  public String getCurrentPageID() {
+    return currentPageID;
+  }
 
-    public void setDisplayName(String displayName) {
-        this.displayName = displayName;
-    }
+  public void setDisplayName(String displayName) {
+    this.displayName = displayName;
+  }
 
-    public void setEmail(String email) {
-        this.email = email;
-    }
+  public void setEmail(String email) {
+    this.email = email;
+  }
 
-    public void setID(String id) {
-        this.id = id;
-    }
+  public void setID(String id) {
+    this.id = id;
+  }
 
-    public void setImageID(String imageID) {
-        this.imageID = imageID;
-    }
+  public void setImageID(String imageID) {
+    this.imageID = imageID;
+  }
 
-    public void setCurrentPageID(String currentPageID) {
-        this.currentPageID = currentPageID;
-    }
+  public void setCurrentPageID(String currentPageID) {
+    this.currentPageID = currentPageID;
+  }
 }
