+++ src/main/webapp/careerquiz.html
@@ -5,14 +5,12 @@
   <meta name = "viewpoint" content="width=device-width, initial-scale=1">
 
   <title>Career Quiz</title>
-
+  
   <link rel="stylesheet" href="style.css">
   <script src="careerquiz.js"></script>
 </head>
 
 <body onload="getQuestionsAndChoices()">
-  <div id="questions-and-choices-div">
-  </div>
-
+  <div id="questions-and-choices-div"></div>
 </body>
 </html>
