+++ src/main/webapp/choosebranch.html
@@ -6,7 +6,5 @@
   <script src="choosebranch.js"></script>
 </head>
 
-<body onload="getRecommendedCareerPath()">
-
-</body>
-
+<body onload="getRecommendedCareerPath()"></body>
+</html>
