+++ src/main/webapp/gameDialogue.js
@@ -1,0 +1,27 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+function getDialogue() {
+  const responsePromise = fetch('/game-dialogue');
+  responsePromise.then(handleResponse);
+}
+
+function handleResponse(response) {
+  const textPromise = response.text();
+  textPromise.then(addDialogueToDom);
+}
+
+function addDialogueToDom(dialogue) {
+  const quoteContainer = document.getElementById('dialogue-container');
+  quoteContainer.innerText = dialogue;
+}
