+++ src/main/webapp/index.html
@@ -7,8 +7,7 @@
   <title>CS Career Explorer</title>
 
   <link href="https://fonts.googleapis.com/css2?family=Press+Start+2P&display=swap" rel="stylesheet">
-  <link rel="stylesheet" href="style.css">
-  
+  <link rel="stylesheet" href="style.css"> 
 </head>
 
 <body>
@@ -18,7 +17,7 @@
     <div class="step-caption">
       <spam class="step-border">
         <h1 class="step-outline">COMPSCI:GO</h1>
-        <a href="#"><button class="step-button">PLAY</button></a>
+        <a href="userAuthPage.html"><button class="step-button">PLAY</button></a>
       </spam>
     </div>
   </div>
@@ -101,10 +100,10 @@
           <a href="https://bit.ly/2Bb4u3H" target="_blank">Java Tutorial for Complete Beginners</a>
         </li>
         <li>
-          <a href="https://bit.ly/3i4nZv8" target="_blank">GeeksforGeeks: Website that teaches you programer with examples and videos</a>
+          <a href="https://bit.ly/3i4nZv8" target="_blank">GeeksforGeeks: Website that teaches you how to code with examples and videos</a>
         </li>
         <li>
-          <a href="https://bit.ly/2Ad8Wyj" target="_blank">YouTube: How start coding videos</a>
+          <a href="https://bit.ly/2Ad8Wyj" target="_blank">YouTube: Videos about How start coding</a>
         </li>
       </ul>
     </div>
@@ -122,7 +121,7 @@
             <a href="https://bit.ly/38b7SYe" target="_blank">CodeWars: Provides a vast amount of problems to practice</a>
           </li>
           <li>
-            <a href="https://bit.ly/3iaG0Ia" target="_blank">YouTube: Videos of Interview to coders</a>
+            <a href="https://bit.ly/3iaG0Ia" target="_blank">YouTube: Coding Interview videos</a>
           </li>
           <li>
             <a href="https://bit.ly/3i9HF0I" target="_blank">How to handle being Interviewed</a>
@@ -135,7 +134,7 @@
       <div id="accordion">
 
         <section id="link1">
-          <a href="#link1">What is a Full-Stack Software Engeneer</a>
+          <a href="#link1">What is a Full-Stack Software Engineer</a>
             <ul class="resource-link">
               <li>
                 <a href="https://bit.ly/2NxlWlg" target="_blank">Definition and Details</a>
