--- src/main/webapp/softwareEngeneer.html
@@ -1,26 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta name = "viewpoint" content="width=device-width, initial-scale=1">
-
-  <title>Software Engineering 101</title>
-
-  <link href="https://fonts.googleapis.com/css2?family=Press+Start+2P&display=swap" rel="stylesheet">
-  <link rel="stylesheet" href="style.css">
-
-  <script src="dialogue.js"></script>
-</head>
-<body onload="getDialogue()">
-  <h2 class="step-outline">Welcome to Software Engineering 001</h2>
-  <img style="float:left;" src="images/teach.png" alt="pixel teacher">
-  <div class="step-wrap, step-dialogue">
-      <div>
-          <p id="dialogue-container"></p>
-      </div>
-      <spam>
-        <button class="step-button" onclick="getDialogue()" name="nextButton">NEXT</button>
-      </spam>
-  </div>
-</body>
-</html>
