+++ src/main/webapp/userAuthPage.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+  <head>
+      <title>Login</title>
+      <script src = "userAuthScript.js"></script>
+      <link rel="stylesheet" href="style.css">
+  </head>
+  <body onload="login()">
+      <h2><u>Login Page</u></h2>
+      <div id = "login-container"></div>
+  </body> 
+</html>
